backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Google AI
google-genai>=1.5.0
google-adk>=1.0.0

# Google Cloud
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.19.0
google-cloud-tasks>=2.16.0
google-cloud-pubsub>=2.27.0
google-cloud-texttospeech>=2.21.0
google-cloud-logging>=3.11.0

# Media
Pillow>=11.0.0
reportlab>=4.2.0
python-pptx>=1.0.0

# WebSocket
websockets>=14.0

# Database
motor>=3.6.0

# Utils
httpx>=0.28.0
aiohttp>=3.11.0
python-multipart>=0.0.19
aiofiles>=24.1.0
